Clean up namespaces, consolidate unit tests, fix Flight compilation
